(DEPRECATED) 1.0.5
Final version. Mod will likely cease to function in v108. This is because KTERA is finally officially using Baldera, and I'd rather deprecate it earlier rather than later.

The mod will still function granted the user manually updates the mod. All you need to do is supply opcodes, if you know how to do that. However, the problem with this is that when we get the Baldera patch, there'll be redundant NPCs, which would be pretty bad looking.